server/requirements.txt
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pip==22.0.2
setuptools==59.6.0
Werkzeug==2.2.2
Flask-SQLAlchemy==3.0.2
simplejson==3.18.0
PyMySQL==1.0.2
python-dotenv==0.21.0
webargs==8.2.0
Flask-SocketIO==5.3.2
eventlet==0.33.2